restructure query code for cli, some test fixes
